Fixed compile warnings on macos

+++ w2/aiUtils.h
@@ -51,7 +51,7 @@ inline void on_closest_enemy_pos(flecs::world &ecs, flecs::entity entity, Callab
 template<typename T>
 inline size_t reg_entity_blackboard_var(flecs::entity entity, const char *bb_name)
 {
-  size_t res = -1;
+  size_t res = size_t(-1);
   entity.set([&](Blackboard &bb)
   {
     res = bb.regName<T>(bb_name);
